src/requirements.txt
# Generated using: pipreqs "C:\Users\esc\Dropbox\Garmin Acrivities Export\Athlete_Data_App_v1.0\src\src_web_p3" --force. 
# --force switch overwrites the existing requirements.txt.

dash_core_components==1.14.1
pytz==2020.5
dash_table==4.11.1
dash_table_experiments==0.6.0
scipy==1.6.0
geopy==2.1.0
dash_html_components==1.1.1
psycopg2==2.8.6
plotly==4.14.3
tzwhere==3.0.3
python_anticaptcha==0.7.1
pandas==1.2.0
dash==1.18.1
fitparse==1.2.0
numpy==1.19.5
dropbox==11.0.0
Flask==1.1.2
func_timeout==4.3.5
python_geoip_python3==1.3
myfitnesspal==1.16.1
mechanize==0.4.5
dicttoxml==1.7.4
psutil==5.8.0
beautifulsoup4==4.9.3
python-geoip-geolite2==2015.0303
pycryptodome==3.9.9
xlrd==2.0.1
